Safari-611.1.10/LayoutTests/http/tests/security/cross-frame-access-object-put-optimization-expected.txt

This tests that you can't get cross-origin object during [[Put]] operation.

PASS data is not "null"
PASS data is not "Cocoa"
PASS data is not "null"
PASS data is not "Cocoa"
PASS: successfullyParsed should be 'true' and is.

TEST COMPLETE
